refactor:타입에 따라 분기처리
