+++ src/components/mypage/UserInfo/PasswordInfo.tsx
@@ -7,7 +7,7 @@ import { passwordChangeSchema, PasswordChangeSchema } from '@/constants/schema/p
 import { useMutation } from '@tanstack/react-query';
 import { patchUserInfo } from '@/api/auth/authAPI';
 import axios from 'axios';
-import CommonInput from '@/components/common/CommonInput';
+import Input from '@/components/common/Input';
 
 interface Props {
   userId: number;
@@ -59,19 +59,19 @@ const PasswordInfo: FC<Props> = ({ userId }) => {
       </div>
       {isEdit ? (
         <div className="userInfo-container-content">
-          <CommonInput
+          <Input
             {...register('newPassword')}
             label="비밀번호"
+            type="password"
             placeholder="새로운 비밀번호를 입력하세요."
             error={errors.newPassword?.message}
-            isVisible
           />
-          <CommonInput
+          <Input
             {...register('newPasswordCheck')}
             label="비밀번호 확인"
+            type="password"
             placeholder="비밀번호를 확인해주세요."
             error={errors.newPasswordCheck?.message}
-            isVisible
           />
         </div>
       ) : null}
